we love programming for other teams yay
Added a simple teleop program for the new robot as well as an autonomous that just drives straight towards the backboard and socres preloaded pixels. I disabled all of your other opmodes for now, but this can be easily reverted by removing the @Disbled at the top of each file.